+++ protocols/ofp12/protocol.go
@@ -2,5 +2,5 @@ package ofp12
 
 const (
 	// Version is the value of version byte in ofp header
-	Version = 0x03
+	Version = uint8(0x03)
 )
