+++ protocols/ofp14/protocol.go
@@ -2,5 +2,5 @@ package ofp14
 
 const (
 	// Version is the value of version byte in ofp header
-	Version = 0x05
+	Version = uint8(0x05)
 )
